debugging-12

+++ src/ru/fizteh/fivt/students/mescherinilya/multifilehashmap/TableProvider.java
@@ -187,6 +187,11 @@ public class TableProvider implements ru.fizteh.fivt.storage.structured.TablePro
 
             }
 
+            File anotherVictim = new File(victimTable.getAbsoluteFile() + File.separator + "signature.tsv");
+            if (anotherVictim.exists() && !anotherVictim.delete()) {
+                throw new IllegalArgumentException("Couldn't delete the file \"signature.tsv\"");
+            }
+
             if (!victimTable.delete()) {
                 throw new IllegalArgumentException("Couldn't delete the directory " + name);
                 //System.err.println("Couldn't delete the directory " + name);
@@ -237,13 +242,7 @@ public class TableProvider implements ru.fizteh.fivt.storage.structured.TablePro
                     } else if (columnType == Float.class) {
                         parsed = Float.parseFloat(core);
                     } else {
-                        for (int j = 0; j < core.length(); ++j) {
-                            if (Character.isWhitespace(core.charAt(j)))
-                                throw new ParseException(
-                                        "There must be no whitespace characters in values!", j);
-                        }
                         parsed = core;
-
                     }
                 } catch (NumberFormatException e) {
                     throw new ParseException("Couldn't parse " + columnType.getName()
